* commit '83d186b3729457864e58973e4cc3e09487e912be':
  [FIX]fix equals problem
  [FIX]fix sonnar issues
  [FIX]Fix problems with creation
  [UPDATE]delete a cascade in a type
  [FIX] fix problem jpa

+++ document-container-service/src/main/java/org/eea/document/service/impl/DocumentServiceImpl.java
@@ -206,17 +206,21 @@ public class DocumentServiceImpl implements DocumentService {
    *
    * @param documentId the document id
    * @param dataFlowId the data flow id
-   *
+   * @param deleteMetabase the delete metabase
    * @throws EEAException the EEA exception
    */
   @Override
   @Modified
   @Async
-  public void deleteDocument(Long documentId, Long dataFlowId) throws EEAException {
+  public void deleteDocument(Long documentId, Long dataFlowId, Boolean deleteMetabase)
+      throws EEAException {
     Session session = null;
     DocumentNodeStore ns = null;
     try {
-      dataflowController.deleteDocument(documentId);
+      if (Boolean.TRUE.equals(deleteMetabase)) {
+        dataflowController.deleteDocument(documentId);
+
+      }
       // Initialize the session
       ns = oakRepositoryUtils.initializeNodeStore();
       Repository repository = oakRepositoryUtils.initializeRepository(ns);
